--- src/tests/Unit/ExampleTest.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Tests\Unit;
-
-use PHPUnit\Framework\TestCase;
-
-class ExampleTest extends TestCase
-{
-    public function test_that_true_is_true(): void
-    {
-        $this->assertTrue(true);
-    }
-}
